refactor: Rename ProjectContextDefinition to ProjectContext and update imports across the codebase
