+++ frontend/src/hooks/useProjects.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ProjectContext, ProjectContextType } from '../context/ProjectContextDefinition';
+import { ProjectContext, ProjectContextType } from '../context/ProjectContext';
 
 export const useProjects = (): ProjectContextType => {
   const context = useContext(ProjectContext);
